cloud_orchestrator/processor/requirements.txt
SQLAlchemy==1.4.14
dependency-injector==4.32.2
pylint==2.9.3
pylint-json2html
transitions==0.8.8
python-terraform==0.10.1
gitpython==3.1.18
aio-pika==6.8.0
PyYAML==5.4.1
psycopg2-binary
cryptography
azure-identity==1.6.1
azure-mgmt-compute==23.0.0
azure-storage-common==2.1.0
azure-storage-blob==12.8.1
azure-data-tables==12.1.0
azure-storage-queue==12.1.6
azure-storage-file-share==12.5.0
azure-core==1.18.0
azure-common==1.1.27
azure-mgmt-resource==20.0.0
azure-keyvault-secrets==4.3.0
msrestazure==0.6.4
azure-mgmt-storage==18.0.0
azure-mgmt-datalake-store==0.5.0
py-grpc-prometheus
jinja2==3.0.1
google-cloud-compute==0.5.0
google-auth==2.1.0
pycryptodome==3.10.1
python-gflags==3.1.2
pika==1.2.0
azure-mgmt-network==19.0.0
azure-mgmt-privatedns==1.0.0
azure-mgmt-dns==8.0.0
pyvmomi==7.0.3
openstacksdk==0.59.0
ansible-base
isi_sdk_8_2_1==0.2.12